move CreateLimitedMemoryStream to StorageTestBase
