+++ sdk/storage/Azure.Storage.Common/tests/Shared/StorageTestBase.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
@@ -14,6 +15,7 @@ using Azure.Core.Pipeline;
 using Azure.Core.Testing;
 using Azure.Identity;
 using Azure.Storage.Sas;
+using Azure.Storage.Tests.Shared;
 using NUnit.Framework;
 
 namespace Azure.Storage.Test.Shared
@@ -465,6 +467,39 @@ namespace Azure.Storage.Test.Shared
                     await Delay(mode, retryDelay);
                 }
             }
+        }
+
+        /// <summary>
+        /// Create a stream of a given size that will not use more than maxMemory memory.
+        /// If the size is greater than maxMemory, a TemporaryFileStream will be used that
+        /// will delete the file in its Dispose method.
+        /// </summary>
+        protected async Task<Stream> CreateLimitedMemoryStream(long size, long maxMemory = Constants.MB * 100)
+        {
+            Stream stream;
+            if (size < maxMemory)
+            {
+                var data = GetRandomBuffer(size);
+                stream = new MemoryStream(data);
+            }
+            else
+            {
+                var path = Path.GetTempFileName();
+                stream = new TemporaryFileStream(path, FileMode.Create);
+                var bufferSize = 4 * Constants.KB;
+
+                while (stream.Position + bufferSize < size)
+                {
+                    await stream.WriteAsync(GetRandomBuffer(bufferSize), 0, bufferSize);
+                }
+                if (stream.Position < size)
+                {
+                    await stream.WriteAsync(GetRandomBuffer(size - stream.Position), 0, (int)(size - stream.Position));
+                }
+                // reset the stream
+                stream.Seek(0, SeekOrigin.Begin);
+            }
+            return stream;
         }
     }
 }
